requirements.txt
alembic==1.0.11
astroid==2.2.5
autopep8==1.4.4
awsebcli==3.15.3
blessed==1.15.0
botocore==1.12.199
cached-property==1.5.1
cement==2.8.2
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.3.9
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
Flask==1.1.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
future==0.16.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.7
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonschema==2.6.0
lazy-object-proxy==1.4.1
Mako==1.0.13
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.0
oauthlib==3.0.2
pathspec==0.5.9
pep8==1.7.1
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pylint==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.2.0
scikit-learn==0.21.3
scipy==1.3.0
semantic-version==2.5.0
six==1.11.0
sklearn==0.0
SQLAlchemy==1.3.5
sqlitedict==1.6.0
termcolor==1.1.0
texttable==0.9.1
TwitterAPI==2.5.9
typed-ast==1.4.0
urllib3==1.24.3
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.4
wrapt==1.11.2
